refs #1089 &Avancement de la structure (Creation d'un noeud ChatMessage)& refs #1616 &Warnings eleves& refs #1091 &Debut d'inplementation d'un chat en command line&

+++ trunk/Commun/Utilitaire/Reseau/PaquetHandlers/PacketHandler.cpp
@@ -6,7 +6,7 @@
 
 HeaderPaquet PacketHandler::handlePacketHeaderReception( PacketReader& pPacketReader )
 {
-    std::string wStringIdentification = std::string((char*)pPacketReader.readString(Paquet::sequenceIdentification.size()+1));
+    std::string wStringIdentification = std::string((char*)pPacketReader.readString((uint32_t) (Paquet::sequenceIdentification.size()+1)));
     if (wStringIdentification != Paquet::sequenceIdentification)
         return HeaderPaquet();
     HeaderPaquet hp;
@@ -20,7 +20,7 @@ void PacketHandler::handlePacketPreparation( Paquet* pPaquet, PacketBuilder& pPa
     pPacketBuilder.includeStringLength(false);
     std::string wType = pPaquet->getOperation();
     pPacketBuilder << pPaquet->sequenceIdentification << wType;
-    pPacketBuilder.addPadding(GestionnaireReseau::longueurMaxOperationReseau - wType.length() - 1);
+    pPacketBuilder.addPadding((uint32_t) (GestionnaireReseau::longueurMaxOperationReseau - wType.length() - 1));
     int wSize = getPacketSize(pPaquet);
     handlePacketPreparationBase(pPaquet, pPacketBuilder, wSize);
     pPacketBuilder.includeStringLength(true);
